Move error boundary code to components level
This makes it appropriate for other components outside of ListResources
to re-use this functionality.
